fix: two clippy lints, and a serial number for the TLS test's certificate
clippy 1.93 flags a cast to the type already there (wow-core, chain.rs) and
first().is_some() for !is_empty() (wow-wallet, refresh.rs).

rcgen will not sign a certificate through a key it does not hold without a
serial number, so the daemon client's TLS test gives it one, as wownerod
gives its generated certificate.

+++ crates/wow-core/src/chain.rs
@@ -521,7 +521,7 @@ impl<D: BlockchainDb> Blockchain<D> {
         if let Some(d) = self.fixed_difficulty {
             return Ok(if self.height() > 0 { d } else { 1 });
         }
-        let count = difficulty_blocks_count(version) as usize;
+        let count = difficulty_blocks_count(version);
         let mut timestamps = Vec::with_capacity(count);
         let mut cumulative = Vec::with_capacity(count);
